refactor: ThemeRepository 인터페이스 분리
