+++ src/main/java/roomescape/repository/ThemeJpaRepository.java
@@ -5,13 +5,11 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.theme.Name;
 import roomescape.domain.theme.Theme;
+import roomescape.domain.theme.ThemeRepository;
 
 @Repository
-public interface ThemeRepository extends JpaRepository<Theme, Long> {
-
-    boolean existsByName(Name name);
+public interface ThemeJpaRepository extends ThemeRepository, JpaRepository<Theme, Long> {
 
     @Query(nativeQuery = true, value = """
         SELECT
