+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.theme.Theme;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.theme.ThemeRepository;
 import roomescape.system.exception.RoomescapeException;
 
 @Service
